requirements.txt
httpx
pillow
python-dotenv
huggingface_hub
hf-xet
groq
opencv-python
numpy
vtracer